requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
motor==3.4.0
orjson==3.10.0
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.10.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
